+++ src/commands/dataconnect-sdk-generate.ts
@@ -20,11 +20,24 @@ export const command = new Command("dataconnect:sdk:generate")
         const cwd = options.cwd || process.cwd();
         configDir = path.resolve(path.join(cwd), configDir);
       }
-      const serviceInfo = await load(projectId, service.location, configDir);
+      const serviceInfo = await load(projectId, configDir);
+      const hasGeneratables = serviceInfo.connectorInfo.some((c) => {
+        return (
+          c.connectorYaml.generate?.javascriptSdk ||
+          c.connectorYaml.generate?.kotlinSdk ||
+          c.connectorYaml.generate?.swiftSdk
+        );
+      });
+      if (!hasGeneratables) {
+        logger.warn("No generated SDKs have been declared in connector.yaml files.");
+        logger.warn(
+          "See https://firebase.google.com/docs/data-connect/quickstart#configure-sdk-outputs for examples of how to configure generated SDKs.",
+        );
+        return;
+      }
       for (const conn of serviceInfo.connectorInfo) {
         const output = await DataConnectEmulator.generate({
           configDir,
-          locationId: service.location,
           connectorId: conn.connectorYaml.connectorId,
         });
         logger.info(output);