+++ src/emulator/controller.ts
@@ -51,7 +51,7 @@ import { Runtime, isRuntime } from "../deploy/functions/runtimes/supported";
 import { AuthEmulator, SingleProjectMode } from "./auth";
 import { DatabaseEmulator, DatabaseEmulatorArgs } from "./databaseEmulator";
 import { EventarcEmulator } from "./eventarcEmulator";
-import { DataConnectEmulator, getLocalConnectionString } from "./dataconnectEmulator";
+import { DataConnectEmulator } from "./dataconnectEmulator";
 import { FirestoreEmulator, FirestoreEmulatorArgs } from "./firestoreEmulator";
 import { HostingEmulator } from "./hostingEmulator";
 import { PubsubEmulator } from "./pubsubEmulator";
@@ -834,19 +834,14 @@ export async function startAll(
       const cwd = options.cwd || process.cwd();
       configDir = path.resolve(path.join(cwd), configDir);
     }
-    const localConnectionString = getLocalConnectionString(options.rc);
     const dataConnectEmulator = new DataConnectEmulator({
       listen: listenForEmulator.dataconnect,
       projectId,
       auto_download: true,
       configDir,
-      locationId: config[0].location,
-      localConnectionString,
+      rc: options.rc,
     });
     await startEmulator(dataConnectEmulator);
-    if (!utils.isVSCodeExtension()) {
-      await dataConnectEmulator.connectToPostgres(localConnectionString);
-    }
   }
 
   if (listenForEmulator.storage) {