+++ src/emulator/downloadableEmulators.ts
@@ -57,20 +57,20 @@ const EMULATOR_UPDATE_DETAILS: { [s in DownloadableEmulators]: EmulatorUpdateDet
   dataconnect:
     process.platform === "darwin"
       ? {
-          version: "1.2.0",
-          expectedSize: 23954240,
-          expectedChecksum: "0f250761959519bb5a28fed76ceab2cb",
+          version: "1.2.2",
+          expectedSize: 24007488,
+          expectedChecksum: "c1fb77895203681479ee5dd22d57249f",
         }
       : process.platform === "win32"
         ? {
-            version: "1.2.0",
-            expectedSize: 24360960,
-            expectedChecksum: "168ce32c742e1d26037c52bdbb7d871c",
+            version: "1.2.2",
+            expectedSize: 24414208,
+            expectedChecksum: "7e263c2b2bc9055ead2db8102e883534",
           }
         : {
-            version: "1.2.0",
-            expectedSize: 23970052,
-            expectedChecksum: "2ca17e4009a9ebae0f7c983bafff2ee6",
+            version: "1.2.2",
+            expectedSize: 24023300,
+            expectedChecksum: "12467418226ac9657fb64b4d719d0e1d",
           },
 };
 
@@ -180,7 +180,7 @@ export const DownloadDetails: { [s in DownloadableEmulators]: EmulatorDownloadDe
       expectedChecksum: EMULATOR_UPDATE_DETAILS.dataconnect.expectedChecksum,
       skipChecksumAndSize: false,
       namePrefix: "dataconnect-emulator",
-      auth: true,
+      auth: false,
     },
   },
 };
@@ -273,7 +273,7 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
     shell: false,
   },
   pubsub: {
-    binary: getExecPath(Emulators.PUBSUB)!,
+    binary: `${getExecPath(Emulators.PUBSUB)!}`,
     args: [],
     optionalArgs: ["port", "host"],
     joinArgs: true,
@@ -287,16 +287,13 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
     shell: false,
   },
   dataconnect: {
-    binary: getExecPath(Emulators.DATACONNECT),
-    args: ["dev"],
+    binary: `${getExecPath(Emulators.DATACONNECT)}`,
+    args: ["--logtostderr", "-v=2", "dev"],
     optionalArgs: [
       "listen",
       "config_dir",
-      "project_id",
-      "service_location",
       "disable_sdk_generation",
       "resolvers_emulator",
-      "vertex_location",
       "rpc_retry_count",
     ],
     joinArgs: true,
@@ -441,6 +438,9 @@ async function _runBinary(
       };
       if (command.shell && utils.IS_WINDOWS) {
         opts.shell = true;
+        if (command.binary.includes(" ")) {
+          command.binary = `"${command.binary}"`;
+        }
       }
       emulator.instance = childProcess.spawn(command.binary, command.args, opts);
     } catch (e: any) {