+++ src/emulator/portUtils.ts
@@ -10,6 +10,7 @@ import { Emulators, ListenSpec } from "./types";
 import { Constants } from "./constants";
 import { EmulatorLogger } from "./emulatorLogger";
 import { execSync } from "node:child_process";
+import { checkIfDataConnectEmulatorRunningOnAddress } from "./dataconnectEmulator";
 
 // See:
 // - https://stackoverflow.com/questions/4313403/why-do-browsers-block-some-ports
@@ -302,6 +303,20 @@ export async function resolveHostAndAssignPorts(
           if (listenable) {
             available.push(listen);
           } else {
+            if (/^dataconnect/i.exec(name)) {
+              const alreadyRunning = await checkIfDataConnectEmulatorRunningOnAddress(listen);
+              // If there is already a running Data Connect emulator on this address, we're gonna try to use it.
+              // If it's for a different service, we'll error out later from DataconnectEmulator.start().
+              if (alreadyRunning) {
+                emuLogger.logLabeled(
+                  "DEBUG",
+                  "dataconnect",
+                  `Detected already running emulator on ${listen.address}:${listen.port}. Will attempt to reuse it.`,
+                );
+              }
+              available.push(listen);
+              continue;
+            }
             if (!portFixed) {
               // Try to find another port to avoid any potential conflict.
               if (i > 0) {