+++ src/test/fixtures/extension-yamls/hello-world/index.ts
@@ -1,0 +1,4 @@
+/**
+ * A valid extension directory containing a full-blown extension.yaml.
+ */
+export const FIXTURE_DIR = __dirname;